+++ odps-sdk/odps-sdk-core/src/main/java/com/aliyun/odps/sqa/ExecuteMode.java
@@ -5,6 +5,7 @@ package com.aliyun.odps.sqa;
  */
 public enum ExecuteMode {
   INTERACTIVE,
+  INTERACTIVE_V2,
   OFFLINE,
   NONE
 }