+++ odps-sdk/odps-sdk-core/src/main/java/com/aliyun/odps/sqa/SQLExecutor.java
@@ -291,6 +291,13 @@ public interface SQLExecutor {
   public boolean isRunningInInteractiveMode();
 
   /**
+   * 判断当前执行的sql使用的模式
+   *
+   * @return INTERACTIVE, INTERACTIVE_V2, OFFLINE, NONE
+   */
+  public ExecuteMode getExecuteMode();
+
+  /**
    * 判断当前 SQLExecutor 是否使用 Instance Tunnel 取结果
    *
    * @return