+++ odps-sdk/odps-sdk-core/src/main/java/com/aliyun/odps/utils/TeaUtils.java
@@ -131,4 +131,11 @@ public class TeaUtils {
       throw new RuntimeException(e.getMessage(), e);
     }
   }
+
+  public static String toString(Object val) {
+    if (val == null) {
+      return "null";
+    }
+    return val.toString();
+  }
 }