+++ odps-sdk/odps-sdk-core/src/test/java/com/aliyun/odps/sqa/SQLExecutorV2Test.java
@@ -1,47 +1,72 @@
 package com.aliyun.odps.sqa;
 
 import java.io.IOException;
-import java.io.OutputStream;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.util.Base64;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 
+import org.junit.Assert;
 import org.junit.BeforeClass;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import com.aliyun.odps.Odps;
 import com.aliyun.odps.OdpsException;
 import com.aliyun.odps.Quota;
-import com.aliyun.odps.account.Account;
-import com.aliyun.odps.account.AliyunAccount;
+import com.aliyun.odps.commons.transport.OdpsTestUtils;
 import com.aliyun.odps.data.Record;
 import com.aliyun.odps.data.ResultSet;
+import com.aliyun.odps.sqa.v2.InfoResultSet;
+import com.aliyun.odps.utils.StringUtils;
 
 /**
  * @author dingxin (zhangdingxin.zdx@alibaba-inc.com)
  */
-@Ignore
 public class SQLExecutorV2Test {
 
   private static SQLExecutor executor;
   private static SQLExecutor tunnelExecutor;
+  private static SQLExecutor commandApiExecutor;
   private static Odps odps;
-  private static final String QUOTA_NAME = "mx_vw";
+  private static final String QUOTA_NAME = "dailyrunnew_mcqa2_quota_nick";
 
   private static Map<String, String> hints;
 
-  private static Odps getTestOdps() throws OdpsException {
-    Account
-        account =
-        new AliyunAccount("ak", "sk");
-    Odps odps = new Odps(account);
-    odps.setEndpoint("endpoint");
-    odps.setDefaultProject("vw2");
-    return odps;
+  @BeforeClass
+  public static void initExecutor() throws OdpsException {
+    SQLExecutorBuilder sqlExecutorBuilder = new SQLExecutorBuilder();
+    odps = OdpsTestUtils.newDefaultOdps();
+    odps.setDefaultProject("odps2_vm");
+    executor =
+        sqlExecutorBuilder.odps(odps)
+            .quotaName(QUOTA_NAME)
+            .useInstanceTunnel(false)
+            .enableMcqaV2(true)
+            .build();
+    tunnelExecutor = new SQLExecutorBuilder().odps(odps)
+        .quotaName(QUOTA_NAME)
+        .useInstanceTunnel(true)
+        .enableMcqaV2(true)
+        .build();
+    commandApiExecutor = new SQLExecutorBuilder().odps(odps)
+        .quotaName(QUOTA_NAME)
+        .enableCommandApi(true)
+        .enableMcqaV2(true)
+        .build();
+    hints = new HashMap<>();
+  }
+
+  @Test
+  public void testIsSelect() throws Exception {
+    executor.run("select 1;", hints);
+    executor.getInstance().waitForTerminatedAndGetResult();
+    boolean select = executor.getInstance().isSelect(executor.getTaskName());
+    Assert.assertTrue(select);
+
+    executor.run("create table if not exists mcqa2_test(c1 string) lifecycle 1;", hints);
+    executor.getInstance().waitForTerminatedAndGetResult();
+    boolean select2 = executor.getInstance().isSelect(executor.getTaskName());
+    Assert.assertFalse(select2);
   }
 
 
@@ -50,20 +75,16 @@ public class SQLExecutorV2Test {
     executor.run("select 1;", hints);
     executor.getResultSet().forEach(System.out::println);
     executor.cancel();
-    try {
-      executor.getInstance().stop();
-    } catch (OdpsException e) {
-      System.out.println(e.getErrorCode());
-      if (e.getErrorCode().equals("InvalidStateSetting")) {
-        System.out.println("good");
-      }
-    }
   }
 
-
   @Test
-  public void testGetTenantId() throws OdpsException {
-    System.out.println(odps.projects().get("three_schema_project").getTenantId());
+  public void testLoadQuota() throws OdpsException {
+    Quota quota = odps.quotas()
+        .getWlmQuota(odps.getDefaultProject(), QUOTA_NAME, null);
+    boolean interactiveQuota = quota.isInteractiveQuota();
+    Assert.assertTrue(interactiveQuota);
+    String mcqaConnHeader = quota.getMcqaConnHeader();
+    Assert.assertNotNull(mcqaConnHeader);
   }
 
   @Test
@@ -83,40 +104,13 @@ public class SQLExecutorV2Test {
 
     tunnelExecutor.run(sql, hints);
     System.out.println(tunnelExecutor.getLogView());
-    ResultSet resultSet = tunnelExecutor.getResultSet();
-
-    resultSet.forEach(record -> System.out.println(record.toString()));
-  }
-
-
-  @BeforeClass
-  public static void initExecutor() throws OdpsException {
-    SQLExecutorBuilder sqlExecutorBuilder = new SQLExecutorBuilder();
-    odps = getTestOdps();
-    executor =
-        sqlExecutorBuilder.odps(odps)
-            .quotaName(QUOTA_NAME)
-            .useInstanceTunnel(false)
-            .enableMcqaV2(true)
-            .enableCommandApi(true)
-            .build();
-    tunnelExecutor = new SQLExecutorBuilder().odps(odps)
-        .quotaName(QUOTA_NAME)
-        .useInstanceTunnel(true)
-        .enableMcqaV2(true)
-        .build();
-    hints = new HashMap<>();
-    hints.put("odps.service.mode", "only");
-    hints.put("odps.sql.enable.enhanced.local.mode", "false");
-    hints.put("odps.sql.runtime.flag.executionengine_CheckJobPlanVersionInSmode", "false");
-  }
-
-  @Test
-  public void testCheckQuota() throws OdpsException {
-    Quota quota = odps.quotas().getWlmQuota(odps.getDefaultProject(), QUOTA_NAME);
-    System.out.println(
-        quota.isInteractiveQuota());
-    System.out.println(new String(Base64.getDecoder().decode(quota.getMcqaConnHeader())));
+    try {
+      ResultSet resultSet = tunnelExecutor.getResultSet();
+      resultSet.forEach(record -> System.out.println(record.toString()));
+      Assert.fail();
+    } catch (Exception e) {
+      System.out.println(e.getMessage());
+    }
   }
 
   @Test
@@ -127,295 +121,71 @@ public class SQLExecutorV2Test {
   }
 
   @Test
-  public void testGetTaskSummary() throws OdpsException, IOException {
-    executor.run("select 1;", hints);
-    System.out.println(executor.getSummary());
+  public void testCommandApi() throws OdpsException, IOException {
+    commandApiExecutor.run("desc mcqa2_test;", hints);
+    List<Record> result = commandApiExecutor.getResult();
+    Assert.assertTrue(result != null && !result.isEmpty());
+
+    ResultSet resultSet = commandApiExecutor.getResultSet();
+    Assert.assertTrue(resultSet != null && resultSet.getRecordCount() == -1);
   }
 
   @Test
-  public void testSync() throws OdpsException, IOException {
+  public void testGetTaskSummary() throws OdpsException, InterruptedException {
+    executor.run("select 1;", hints);
+    executor.getInstance().waitForSuccess();
+    TimeUnit.SECONDS.sleep(10);
+    String summary = executor.getSummary();
+    System.out.println(summary);
+    Assert.assertTrue(StringUtils.isNotBlank(summary));
+  }
+
+  @Test
+  public void testSelectNonTunnel() throws OdpsException, IOException {
     executor.run("select 1;", hints);
     List<Record> result = executor.getResult();
-    result.forEach(r -> System.out.println(r.toString()));
-
-    System.out.println(executor.getSummary());
+    Assert.assertTrue(result != null && !result.isEmpty());
+    ResultSet resultSet = executor.getResultSet();
+    Assert.assertTrue(resultSet != null && resultSet.getRecordCount() > 0);
   }
 
   @Test
-  public void testSyncByTunnel() throws OdpsException, IOException {
+  public void testNonSelectNonTunnel() throws OdpsException, IOException {
+    executor.run("desc mcqa2_test;", hints);
+    List<Record> result = executor.getResult();
+    Assert.assertTrue(result != null && !result.isEmpty());
+    result.forEach(r -> System.out.println(r.toString()));
+
+    ResultSet resultSet = executor.getResultSet();
+    Assert.assertTrue(resultSet instanceof InfoResultSet);
+    Assert.assertTrue(resultSet.getRecordCount() == 1);
+  }
+
+
+  @Test
+  public void testSelectByTunnel() throws OdpsException, IOException {
     tunnelExecutor.run("select 1;", hints);
     List<Record> result = tunnelExecutor.getResult();
-    result.forEach(r -> System.out.println(r.toString()));
-
-    System.out.println(tunnelExecutor.getSummary());
+    Assert.assertTrue(result != null && !result.isEmpty());
+    ResultSet resultSet = tunnelExecutor.getResultSet();
+    Assert.assertTrue(resultSet != null && resultSet.getRecordCount() > 0);
   }
 
   @Test
-  public void testTpcdsQ4() throws OdpsException, IOException {
-    executor.run("SELECT  dt.d_year\n"
-                 + "        ,item.i_brand_id brand_id\n"
-                 + "        ,item.i_brand brand\n"
-                 + "        ,SUM(ss_sales_price) sum_agg\n"
-                 + "FROM    date_dim dt\n"
-                 + ",       store_sales\n"
-                 + ",       item\n"
-                 + "WHERE   dt.d_date_sk = store_sales.ss_sold_date_sk\n"
-                 + "AND     store_sales.ss_item_sk = item.i_item_sk\n"
-                 + "AND     item.i_manufact_id = 190\n"
-                 + "AND     dt.d_moy = 12\n"
-                 + "GROUP BY dt.d_year\n"
-                 + "          ,item.i_brand\n"
-                 + "          ,item.i_brand_id\n"
-                 + "ORDER BY dt.d_year,sum_agg DESC,brand_id\n"
-                 + "LIMIT   100\n"
-                 + ";", hints);
-    ResultSet resultSet = executor.getResultSet();
-    resultSet.forEach(System.out::println);
-    System.out.println(resultSet.getRecordCount());
-    System.out.println(executor.getLogView());
-
-    //System.out.println(executor.getInstance().getTaskDetailJson2(executor.getTaskName()));
-
-    System.out.println(executor.getSummary());
-  }
-
-  @Test
-  public void testTpcdsQ4UseTunnel() throws OdpsException, IOException {
-    tunnelExecutor.run("SELECT  dt.d_year\n"
-                 + "        ,item.i_brand_id brand_id\n"
-                 + "        ,item.i_brand brand\n"
-                 + "        ,SUM(ss_sales_price) sum_agg\n"
-                 + "FROM    date_dim dt\n"
-                 + ",       store_sales\n"
-                 + ",       item\n"
-                 + "WHERE   dt.d_date_sk = store_sales.ss_sold_date_sk\n"
-                 + "AND     store_sales.ss_item_sk = item.i_item_sk\n"
-                 + "AND     item.i_manufact_id = 190\n"
-                 + "AND     dt.d_moy = 12\n"
-                 + "GROUP BY dt.d_year\n"
-                 + "          ,item.i_brand\n"
-                 + "          ,item.i_brand_id\n"
-                 + "ORDER BY dt.d_year,sum_agg DESC,brand_id\n"
-                 + "LIMIT   100\n"
-                 + ";", hints);
-    ResultSet resultSet = tunnelExecutor.getResultSet(10L,10L, null);
-    resultSet.forEach(System.out::println);
-    System.out.println(resultSet.getRecordCount());
+  public void testNonSelectByTunnel() throws OdpsException, IOException {
+    tunnelExecutor.run("desc mcqa2_test;", hints);
+    List<Record> result = tunnelExecutor.getResult();
+    Assert.assertTrue(result != null && !result.isEmpty());
+    ResultSet resultSet = tunnelExecutor.getResultSet();
+    Assert.assertTrue(resultSet instanceof InfoResultSet);
+    Assert.assertTrue(resultSet.getRecordCount() == 1);
   }
 
   @Test
   public void generateLogview() throws OdpsException {
-    try {
-      executor.run("select 1;", hints);
-      System.out.println(executor.getResult());
-    } catch (Exception e) {
-
-    }
-    System.out.println(executor.getLogView());
-  }
-
-  @Test
-  public void testReadResultFromInstanceTunnel() throws OdpsException, IOException {
-
-    tunnelExecutor.run("select 1;", hints);
-    System.out.println(tunnelExecutor.getResult());
-
-  }
-
-  @Test
-  public void testTpcdsQuery72() throws OdpsException, IOException {
-    String tpcdsQuery = "SELECT  i_item_desc\n"
-               + "        ,w_warehouse_name\n"
-               + "        ,d1.d_week_seq\n"
-               + "        ,SUM(CASE WHEN p_promo_sk IS NULL THEN 1 ELSE 0 END) no_promo\n"
-               + "        ,SUM(CASE WHEN p_promo_sk IS NOT NULL THEN 1 ELSE 0 END) promo\n"
-               + "        ,COUNT(*) total_cnt\n"
-               + "FROM    catalog_sales\n"
-               + "JOIN    inventory\n"
-               + "ON      (\n"
-               + "            cs_item_sk = inv_item_sk\n"
-               + ")\n"
-               + "JOIN    warehouse\n"
-               + "ON      (\n"
-               + "            w_warehouse_sk = inv_warehouse_sk\n"
-               + ")\n"
-               + "JOIN    item\n"
-               + "ON      (\n"
-               + "            i_item_sk = cs_item_sk\n"
-               + ")\n"
-               + "JOIN    customer_demographics\n"
-               + "ON      (\n"
-               + "            cs_bill_cdemo_sk = cd_demo_sk\n"
-               + ")\n"
-               + "JOIN    household_demographics\n"
-               + "ON      (\n"
-               + "            cs_bill_hdemo_sk = hd_demo_sk\n"
-               + ")\n"
-               + "JOIN    date_dim d1\n"
-               + "ON      (\n"
-               + "            cs_sold_date_sk = d1.d_date_sk\n"
-               + ")\n"
-               + "JOIN    date_dim d2\n"
-               + "ON      (\n"
-               + "            inv_date_sk = d2.d_date_sk\n"
-               + ")\n"
-               + "JOIN    date_dim d3\n"
-               + "ON      (\n"
-               + "            cs_ship_date_sk = d3.d_date_sk\n"
-               + ")\n"
-               + "LEFT OUTER JOIN promotion\n"
-               + "ON      (\n"
-               + "            cs_promo_sk = p_promo_sk\n"
-               + ")\n"
-               + "LEFT OUTER JOIN catalog_returns\n"
-               + "ON      (\n"
-               + "            cr_item_sk = cs_item_sk\n"
-               + "            AND     cr_order_number = cs_order_number\n"
-               + ")\n"
-               + "WHERE   d1.d_week_seq = d2.d_week_seq\n"
-               + "AND     inv_quantity_on_hand < cs_quantity\n"
-               + "AND     d3.d_date > d1.d_date + 5\n"
-               + "AND     hd_buy_potential = '1001-5000'\n"
-               + "AND     d1.d_year = 1999\n"
-               + "AND     cd_marital_status = 'W'\n"
-               + "GROUP BY i_item_desc\n"
-               + "          ,w_warehouse_name\n"
-               + "          ,d1.d_week_seq\n"
-               + "ORDER BY total_cnt DESC,i_item_desc,w_warehouse_name,d_week_seq\n"
-               + "LIMIT   100\n"
-               + ";";
-    System.out.println(tpcdsQuery.replace("\n", " "));
-    //System.out.println(tpcdsQuery);
-    executor.run(tpcdsQuery, hints);
-    System.out.println(executor.getLogView());
-    ResultSet resultSet = executor.getResultSet();
-    resultSet.forEach(System.out::println);
-    System.out.println(resultSet.getRecordCount());
-  }
-
-  public static void main(String[] args) {
-    String urlString = "http://11.158.199.208:9991/mcqa/projects/tpcds_10g_test3/instances?curr_project=tpcds_10g_test3";
-
-    // 新的 TPC-DS 查询
-    String tpcdsQuery = "SELECT  i_item_desc\n"
-                        + "        ,w_warehouse_name\n"
-                        + "        ,d1.d_week_seq\n"
-                        + "        ,SUM(CASE WHEN p_promo_sk IS NULL THEN 1 ELSE 0 END) no_promo\n"
-                        + "        ,SUM(CASE WHEN p_promo_sk IS NOT NULL THEN 1 ELSE 0 END) promo\n"
-                        + "        ,COUNT(*) total_cnt\n"
-                        + "FROM    catalog_sales\n"
-                        + "JOIN    inventory\n"
-                        + "ON      (\n"
-                        + "            cs_item_sk = inv_item_sk\n"
-                        + ")\n"
-                        + "JOIN    warehouse\n"
-                        + "ON      (\n"
-                        + "            w_warehouse_sk = inv_warehouse_sk\n"
-                        + ")\n"
-                        + "JOIN    item\n"
-                        + "ON      (\n"
-                        + "            i_item_sk = cs_item_sk\n"
-                        + ")\n"
-                        + "JOIN    customer_demographics\n"
-                        + "ON      (\n"
-                        + "            cs_bill_cdemo_sk = cd_demo_sk\n"
-                        + ")\n"
-                        + "JOIN    household_demographics\n"
-                        + "ON      (\n"
-                        + "            cs_bill_hdemo_sk = hd_demo_sk\n"
-                        + ")\n"
-                        + "JOIN    date_dim d1\n"
-                        + "ON      (\n"
-                        + "            cs_sold_date_sk = d1.d_date_sk\n"
-                        + ")\n"
-                        + "JOIN    date_dim d2\n"
-                        + "ON      (\n"
-                        + "            inv_date_sk = d2.d_date_sk\n"
-                        + ")\n"
-                        + "JOIN    date_dim d3\n"
-                        + "ON      (\n"
-                        + "            cs_ship_date_sk = d3.d_date_sk\n"
-                        + ")\n"
-                        + "LEFT OUTER JOIN promotion\n"
-                        + "ON      (\n"
-                        + "            cs_promo_sk = p_promo_sk\n"
-                        + ")\n"
-                        + "LEFT OUTER JOIN catalog_returns\n"
-                        + "ON      (\n"
-                        + "            cr_item_sk = cs_item_sk\n"
-                        + "            AND     cr_order_number = cs_order_number\n"
-                        + ")\n"
-                        + "WHERE   d1.d_week_seq = d2.d_week_seq\n"
-                        + "AND     inv_quantity_on_hand < cs_quantity\n"
-                        + "AND     d3.d_date > d1.d_date + 5\n"
-                        + "AND     hd_buy_potential = '1001-5000'\n"
-                        + "AND     d1.d_year = 1999\n"
-                        + "AND     cd_marital_status = 'W'\n"
-                        + "GROUP BY i_item_desc\n"
-                        + "          ,w_warehouse_name\n"
-                        + "          ,d1.d_week_seq\n"
-                        + "ORDER BY total_cnt DESC,i_item_desc,w_warehouse_name,d_week_seq\n"
-                        + "LIMIT   100\n"
-                        + ";";
-    String xmlData = "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"yes\"?>" +
-                     "<Instance><Job><Priority>9</Priority><Tasks>" +
-                     "<SQL><Name>test_sql_task_112</Name><Config>" +
-                     "<Property><Name>settings</Name><Value>{\"odps.service.mode\":\"all\"," +
-                     "\"odps.sql.runtime.flag.executionengine_CheckJobPlanVersionInSmode\":\"false\"," +
-                     "\"odps.sql.select.output.format\":\"HumanReadable\"}</Value></Property>" +
-                     "</Config><Query>" + escapeXml(tpcdsQuery) + "</Query></SQL>" +
-                     "</Tasks></Job></Instance>";
-
-    HttpURLConnection connection = null;
-
-    try {
-      // 创建 URL 对象
-      URL url = new URL(urlString);
-      // 打开连接
-      connection = (HttpURLConnection) url.openConnection();
-      // 设置请求方法为 POST
-      connection.setRequestMethod("POST");
-      // 设置请求头
-      connection.setRequestProperty("x-odps-request-account", "1365937150772213");
-      // 指定请求体的类型
-      connection.setRequestProperty("Content-Type", "application/xml");
-      // 启用输入输出流
-      connection.setDoOutput(true);
-
-
-      connection.setDoInput(true);
-
-
-      // 发送请求体
-      try (OutputStream os = connection.getOutputStream()) {
-        byte[] input = xmlData.getBytes("utf-8");
-        os.write(input, 0, input.length);
-      }
-
-      // 检查响应码
-      int responseCode = connection.getResponseCode();
-      System.out.println("Response Code: " + responseCode);
-      // 打印所有响应 headers
-      Map<String, List<String>> headerFields = connection.getHeaderFields();
-      for (Map.Entry<String, List<String>> header : headerFields.entrySet()) {
-        // 打印 header 名字和对应的值
-        System.out.println(header.getKey() + ": " + String.join(", ", header.getValue()));
-      }
-
-    } catch (Exception e) {
-      e.printStackTrace();
-    } finally {
-      if (connection != null) {
-        connection.disconnect();
-      }
-    }
-  }
-
-  private static String escapeXml(String xml) {
-    return xml.replace("&", "&amp;")
-        .replace("<", "&lt;")
-        .replace(">", "&gt;")
-        .replace("\"", "&quot;")
-        .replace("'", "&apos;");
+    executor.run("select 1;", hints);
+    String logView = executor.getLogView();
+    System.out.println(logView);
+    Assert.assertNotNull(logView);
   }
 }